DD-659. deposit-to-dataverse must use pre-staged.csv to get checksums of pruned files
