+++ src/test/scala/nl.knaw.dans.easy.dd2d/mapping/LanguageSpec.scala
@@ -17,17 +17,17 @@ package nl.knaw.dans.easy.dd2d.mapping
 
 import better.files.File
 import nl.knaw.dans.easy.dd2d.mapping.Language.toCitationBlockLanguage
-import nl.knaw.dans.easy.dd2d.{ Configuration, TestSupportFixture }
+import nl.knaw.dans.easy.dd2d.{ Configuration, TestSupportFixture, loadCsvToMap }
 
 import java.nio.file.Paths
 
 class LanguageSpec extends TestSupportFixture {
-  private val iso1ToDataverseLanguage = Configuration
-    .loadCsvToMap(File(Paths.get("src/main/assembly/dist/install/iso639-1-to-dv.csv").toAbsolutePath),
+  private val iso1ToDataverseLanguage =
+    loadCsvToMap(File(Paths.get("src/main/assembly/dist/install/iso639-1-to-dv.csv").toAbsolutePath),
       keyColumn = "ISO639-1",
       valueColumn = "Dataverse-language").get
-  private val iso2ToDataverseLanguage = Configuration
-    .loadCsvToMap(File(Paths.get("src/main/assembly/dist/install/iso639-2-to-dv.csv").toAbsolutePath),
+  private val iso2ToDataverseLanguage =
+    loadCsvToMap(File(Paths.get("src/main/assembly/dist/install/iso639-2-to-dv.csv").toAbsolutePath),
       keyColumn = "ISO639-2",
       valueColumn = "Dataverse-language").get
